+++ test/python/echo_text_with_newline_control.py
@@ -1,0 +1,5 @@
+import sys, json, os
+
+for line in sys.stdin:
+  line = line.replace('$', os.linesep)
+  print(line[:-1], end='')